requirements.txt
openpyxl
python-barcode
Pillow
streamlit
reportlab